+++ src/main/java/com/mtvs/quizlog/domain/like/dto/TeacherLikeRankingDto.java
@@ -1,0 +1,3 @@
+package com.mtvs.quizlog.domain.like.dto;
+
+public record TeacherLikeRankingDto(String nickname, Long likeCount) { }
