+++ src/main/resources/templates/main.html
@@ -19,7 +19,7 @@
             <div></div>
         </div>
         <div class="main-content">
-            <h3>기타어쩌구</h3>
+            <h3>선생님 랭킹 (좋아요 순)</h3>
             <div></div>
         </div>
     </div>
