+++ src/main/resources/templates/test.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <title>메인 페이지</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
+</head>
+<body>
+
+<div class="container mt-5">
+    <h2>🔥 좋아요 Top 5 선생님</h2>
+
+    <div id="teacherCarousel" class="carousel slide" data-bs-ride="carousel">
+        <div class="carousel-inner">
+            <div class="carousel-item"
+                 th:each="teacher, iterStat : ${topTeachers}"
+                 th:classappend="${iterStat.index == 0} ? 'active'">
+                <h4 class="text-center" th:text="${teacher.nickname}"></h4>
+            </div>
+
+            <!-- 리스트 비었을 때 대체 뷰 -->
+            <div class="carousel-item active" th:if="${topTeachers == null or topTeachers.size() == 0}">
+                <h4 class="text-center">표시할 선생님이 없습니다.</h4>
+            </div>
+        </div>
+
+
+        <button class="carousel-control-prev" type="button" data-bs-target="#teacherCarousel" data-bs-slide="prev">
+            <span class="carousel-control-prev-icon"></span>
+        </button>
+        <button class="carousel-control-next" type="button" data-bs-target="#teacherCarousel" data-bs-slide="next">
+            <span class="carousel-control-next-icon"></span>
+        </button>
+    </div>
+</div>
+
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+</body>
+</html>
